% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Run covr and potentially save results to disk}
\usage{
add_coverage(pkg_path, out_dir)
}
\arguments{
\item{pkg_path}{package installation directory}

\item{out_dir}{directory for saving results}
}
\description{
Run covr and potentially save results to disk
}
\keyword{internal}
