% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{add_rcmdcheck}
\alias{add_rcmdcheck}
\title{Run R CMD CHECK}
\usage{
add_rcmdcheck(out_dir, rcmdcheck_args)
}
\arguments{
\item{out_dir}{directory for saving results}

\item{rcmdcheck_args}{list of arguments to pass to \code{rcmdcheck::rcmdcheck}}
}
\description{
Run R CMD CHECK
}
\details{
rcmdcheck takes either a tarball or an installation directory.

The basename of \code{out_dir} should be the package name and version pasted together
}
\keyword{internal}
