% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard-summary.R
\name{build_risk_summary}
\alias{build_risk_summary}
\title{Summarise overall risks for each package}
\usage{
build_risk_summary(result_dirs, risk_breaks, out_dir, append_out_dir = FALSE)
}
\arguments{
\item{result_dirs}{A vector of output directories}

\item{risk_breaks}{A numeric vector of length 2, with both numbers being
between 0 and 1. These are used for the "breaks" when classifying scores
into Risk categories. For example, for the default value of \code{c(0.3, 0.7)},
for a given score: \verb{0 <= score < 0.3} is "Low Risk", \verb{0.3 <= score < 0.7}
is "Medium Risk", and \verb{0.7 <= score < 1} is "High Risk".}

\item{out_dir}{Output directory for saving scorecard summary. If \code{NULL}, assumes all \code{result_dirs} point to the same output directory}

\item{append_out_dir}{Logical (T/F). If \code{TRUE}, keep the package-specific output directory as part of the returned table.}
}
\description{
Summarise overall risks for each package
}
\keyword{internal}
