% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-overall-scores.R
\name{calc_overall_scores}
\alias{calc_overall_scores}
\title{Calculate overall category scores}
\usage{
calc_overall_scores(scorelist)
}
\description{
This is where we'll need to build the algorithm for weighting
different metrics, etc. We also need to consider if we want
this algorithm (i.e. the weights) to be configurable be the user...
}
