% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{check_missing_artifact}
\alias{check_missing_artifact}
\title{Check for missing links in a traceability matrix}
\usage{
check_missing_artifact(
  exports_df,
  pkg_source_path,
  artifact = c("code_file", "documentation", "test_files")
)
}
\arguments{
\item{exports_df}{data.frame with a column, named \code{exported_function},
containing the names of all exported functions. Can also have other columns
(which will be returned unmodified).}

\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}

\item{artifact}{either \code{"code_file"}, \code{"documentation"}, or \code{"test_files"}.
Must be a column present in \code{exports_df}.}
}
\description{
Check for missing links in a traceability matrix
}
\keyword{internal}
