% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{check_scores_numeric}
\alias{check_scores_numeric}
\title{Check category scores for non-numeric values}
\usage{
check_scores_numeric(pkg_scores, json_path)
}
\arguments{
\item{pkg_scores}{a named list containing the build up of score elements and
overall category scores.}

\item{json_path}{a JSON file path.}
}
\description{
Check category scores for non-numeric values
}
\details{
\code{covr} scores are the only individual scores that allow for NA scores.
However when determining overall category scores, all values should be
coerced to numeric (0 in the case of \code{covr}). Any potential errors or
character values should be coerced to \code{NA} when passed through \code{mean()} in
\code{calc_overall_scores}.

This function serves to confirm that all category scores are in fact numeric
values. The only way a non-numeric value could be anything \emph{other than} \code{NA},
is if \code{mean()} returned an error message or some other value. This scenario
would likely cause issues before writing the score list out to json, so the
additional \code{is.numeric} check is primarily a precautionary measure.
}
\keyword{internal}
