% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{check_scores_valid}
\alias{check_scores_valid}
\title{Check that package scores are valid}
\usage{
check_scores_valid(pkg_scores, json_path)
}
\arguments{
\item{pkg_scores}{a named list containing the build up of score elements and
overall category scores.}

\item{json_path}{a JSON file path.}
}
\description{
Check that individual scores are retained and category scores are numeric.
}
\details{
\strong{Individual Score checks}

The only checks performed on the individual scores are that they still exist
when written out to json. This would only \emph{not} be the case if an individual
score was set to \code{NULL}, removing it from the list.

\strong{Category Score checks}

\code{covr} scores are the only individual scores that allow for NA scores.
However when determining overall category scores, all values should be
coerced to numeric (0 in the case of \code{covr}). Any potential errors or
character values should be coerced to \code{NA} when passed through \code{mean()} in
\code{calc_overall_scores}.

This function serves to confirm that all category scores are in fact numeric
values. The only way a non-numeric value could be anything \emph{other than} \code{NA},
is if \code{mean()} returned an error message or some other value. This scenario
would likely cause issues before writing the score list out to json, so the
additional \code{is.numeric} check is primarily a precautionary measure.
}
\keyword{internal}
