% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-extra-notes.R
\name{create_extra_notes}
\alias{create_extra_notes}
\title{Create extra notes summarizing the covr & rcmdcheck outputs, and documentation}
\usage{
create_extra_notes(results_dir)
}
\arguments{
\item{results_dir}{directory containing json file and individual results. Output file path from \code{\link[=score_pkg]{score_pkg()}}}
}
\description{
Create extra notes summarizing the covr & rcmdcheck outputs, and documentation
}
\keyword{internal}
