% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{create_pkg_env}
\alias{create_pkg_env}
\title{Create environment for sourcing package functions}
\usage{
create_pkg_env(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\value{
an environment
}
\description{
Create environment for sourcing package functions
}
\details{
Inpsired from \code{pkload:::create_ns_env} and \code{methods::setPackageName}
}
\keyword{internal}
