% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{create_score_list_from_riskmetric}
\alias{create_score_list_from_riskmetric}
\title{Format riskmetric results into scorecard list}
\usage{
create_score_list_from_riskmetric(res, pkg_source_path)
}
\arguments{
\item{res}{a named list containing the initial build up of score elements}

\item{pkg_source_path}{path to package source code (untarred)}
}
\description{
Format riskmetric results into scorecard list
}
\keyword{internal}
