% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{find_export_script}
\alias{find_export_script}
\title{Get all exported functions and map them to R script where they are defined}
\usage{
find_export_script(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\description{
Get all exported functions and map them to R script where they are defined
}
\keyword{internal}
