% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-extra-notes.R
\name{find_function_files}
\alias{find_function_files}
\title{Search for functions in a directory}
\usage{
find_function_files(funcs, search_dir, func_declaration = TRUE)
}
\arguments{
\item{funcs}{a vector of function names to search for}

\item{search_dir}{a directory to search for the functions in}

\item{func_declaration}{Logical (T/F). If \code{TRUE}, only look for where the function is actually defined}
}
\description{
Search for functions in a directory
}
\keyword{internal}
