% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{flextable_formatted}
\alias{flextable_formatted}
\title{Autofit and format flextables}
\usage{
flextable_formatted(
  tab,
  autofit = TRUE,
  pg_width = 5,
  column_width = NULL,
  doc_type = "PDF",
  as_flextable = TRUE,
  digits = NULL,
  ...
)
}
\arguments{
\item{tab}{a dataframe or flextable object. Must be coercible to a dataframe}

\item{autofit}{logical (T/F). Whether or not to autofit the table.}

\item{pg_width}{width (in inches) of the table. Generally 1 inch less than the default word document (8 in.)}

\item{column_width}{named vector, where the column names are assigned to the desired \emph{relative} width.
If specified, set these column widths before fitting to word document}

\item{doc_type}{Word, PDF, or HTML. Controls font size}

\item{as_flextable}{logical (T/F). if \code{TRUE}, use \code{as_flextable} instead of \code{flextable}, which has different args and is necessary for grouped data}

\item{digits}{numeric. Number of digits to round to. If \code{NULL}, and \code{as_flextable = TRUE}, flextable will round to one digit.}

\item{...}{additional args to be passed to \code{as_flextable} or \code{flextable}}
}
\value{
a formatted flextable
}
\description{
Autofit and format flextables
}
\details{
column_width is specified using the following convention:

\if{html}{\out{<div class="sourceCode">}}\preformatted{tab \%>\%
flextable_word(column_width = c("col1" = 2, "col2" = 3))
}\if{html}{\out{</div>}}

\code{flextable} by default will make the tables as wide as possible in word. This function will correct the \code{autofit()} feature and make the contents fit.
Other formatting corrections are done depending on the specified \code{doc_type}
}
\keyword{internal}
