% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_colnames_to_title}
\alias{format_colnames_to_title}
\title{Format column names to be user friendly}
\usage{
format_colnames_to_title(df)
}
\arguments{
\item{df}{a dataframe}
}
\description{
Replaces underscores (\verb{_}) with spaces and makes the first letter capitalized
}
\keyword{internal}
