% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_comments}
\alias{format_comments}
\title{Format vector of comments text}
\usage{
format_comments(comments_block)
}
\arguments{
\item{comments_block}{character vector of comments text. Can include bullets}
}
\description{
Format vector of comments text
}
\keyword{internal}
