% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_metadata}
\alias{format_metadata}
\title{Format system info and metadata}
\usage{
format_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{metadata list returned by \code{formatted_pkg_scores$metadata}}
}
\value{
a formatted flextable object
}
\description{
Format system info and metadata
}
\keyword{internal}
