% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_mitigation}
\alias{format_mitigation}
\title{Format vector of mitigation text}
\usage{
format_mitigation(mitigation_block)
}
\arguments{
\item{mitigation_block}{character vector of mitigation text. Can include bullets}
}
\description{
Format vector of mitigation text
}
\keyword{internal}
