% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_overall_scores}
\alias{format_overall_scores}
\title{Format Overall Scores}
\usage{
format_overall_scores(formatted_pkg_scores, digits = 2)
}
\arguments{
\item{formatted_pkg_scores}{list containing all scores}

\item{digits}{number of digits to round scores to.}
}
\value{
a formatted flextable object
}
\description{
Format Overall Scores
}
\keyword{internal}
