% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_package_details}
\alias{format_package_details}
\title{Format Package Details}
\usage{
format_package_details(formatted_pkg_scores)
}
\arguments{
\item{formatted_pkg_scores}{list containing all scores}
}
\value{
a formatted flextable object
}
\description{
Formats Documentation, Maintenance and Transparency scores
}
\keyword{internal}
