% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_score_summaries}
\alias{format_score_summaries}
\title{Format summary of overall risks}
\usage{
format_score_summaries(risk_summary_df, digits = 2)
}
\arguments{
\item{risk_summary_df}{summary dataframe containing overall scores, risks,
and package info for each package}

\item{digits}{number of digits to round scores to}
}
\value{
a formatted flextable object
}
\description{
Format summary of overall risks
}
\keyword{internal}
