% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{format_scorecard_version}
\alias{format_scorecard_version}
\title{Format mpn.scorecard version as a footer note}
\usage{
format_scorecard_version(json_ver = NULL, scorecard_ver)
}
\arguments{
\item{json_ver}{version of mpn.scorecard as recorded in \code{score_pkg}.
Set to \code{NULL} for use with \code{render_scorecard_summary}.}

\item{scorecard_ver}{version of mpn.scorecard as recorded in \code{render_scorecard}
or \code{render_scorecard_summary}}
}
\description{
Formats mpn.scorecard version as a footer note for use in \code{render_scorecard}
or \code{render_scorecard_summary}
}
\keyword{internal}
