% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{format_scores_for_render}
\alias{format_scores_for_render}
\title{Prepare the raw scores to be rendered into PDF}
\usage{
format_scores_for_render(pkg_scores, risk_breaks)
}
\description{
Prepare the raw scores to be rendered into PDF
}
