% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_testing_scores}
\alias{format_testing_scores}
\title{Format testing scores}
\usage{
format_testing_scores(formatted_pkg_scores)
}
\arguments{
\item{formatted_pkg_scores}{list containing all scores}
}
\value{
a formatted flextable object
}
\description{
Format testing scores
}
\keyword{internal}
