% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{get_all_functions}
\alias{get_all_functions}
\title{list all functions defined in the package code}
\usage{
get_all_functions(pkg_source_path)
}
\value{
A data.frame with the columns \code{func} and \code{code_file} with a row for
every function defined in the package.
}
\description{
list all functions defined in the package code
}
\keyword{internal}
