% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{get_exports}
\alias{get_exports}
\title{list all package exports}
\usage{
get_exports(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\description{
list all package exports
}
\keyword{internal}
