% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-extra-notes.R
\name{get_exports_documented}
\alias{get_exports_documented}
\title{Get aliases of exported functions}
\usage{
get_exports_documented(pkg_tar_path)
}
\arguments{
\item{pkg_tar_path}{path to a tarball}
}
\value{
a tibble
}
\description{
Returns the aliases of exported functions, the name of the \code{man} file, and the R script the function is contained in
}
\keyword{internal}
