% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{get_flex_caption}
\alias{get_flex_caption}
\title{Create a flextable caption colored by risk}
\usage{
get_flex_caption(formatted_pkg_scores, category)
}
\arguments{
\item{formatted_pkg_scores}{list containing all scores}

\item{category}{one of \code{c('testing', 'documentation', 'maintenance', 'transparency', 'overall')}}
}
\value{
a flextable caption
}
\description{
Create a flextable caption colored by risk
}
\keyword{internal}
