% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{get_overall_labels}
\alias{get_overall_labels}
\title{Retrieve and format category labels}
\usage{
get_overall_labels(formatted_pkg_scores, category)
}
\arguments{
\item{formatted_pkg_scores}{list containing all scores}

\item{category}{one of \code{c('testing', 'documentation', 'maintenance', 'transparency', 'overall')}}
}
\value{
a character string in the format of \if{html}{\out{<category>}}: \if{html}{\out{<risk>}}
}
\description{
Retrieve and format category labels
}
\keyword{internal}
