% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pkg_desc}
\alias{get_pkg_desc}
\title{Read Description file and parse the package name and version}
\usage{
get_pkg_desc(pkg_path, fields = NULL)
}
\arguments{
\item{pkg_path}{path to installed/untarred package}
}
\description{
Read Description file and parse the package name and version
}
\keyword{internal}
