% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{get_tests}
\alias{get_tests}
\title{Tabulate all tests in a package directory}
\usage{
get_tests(pkg_source_path, test_dirs)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}

\item{test_dirs}{one or more directories containing test scripts}
}
\value{
a dataframe containing all test names and the test file they are associated with
}
\description{
Tabulate all tests in a package directory
}
\keyword{internal}
