% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-extra-notes.R
\name{make_traceability_matrix}
\alias{make_traceability_matrix}
\title{Get aliases of exported functions}
\usage{
make_traceability_matrix(pkg_tar_path, results_dir)
}
\arguments{
\item{pkg_tar_path}{path to a tarball}

\item{results_dir}{directory containing json file and individual results. Output file path from \code{\link[=score_pkg]{score_pkg()}}}
}
\value{
a tibble
}
\description{
Returns the aliases of exported functions, the name of the \code{man} file, and the R script the function is contained in
}
\keyword{internal}
