% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{map_functions_to_scripts}
\alias{map_functions_to_scripts}
\title{Get all exported functions and map them to R script where they are defined}
\usage{
map_functions_to_scripts(exports_df, pkg_source_path, verbose = FALSE)
}
\arguments{
\item{exports_df}{data.frame with a column, named \code{exported_function},
containing the names of all exported functions. Can also have other columns
(which will be returned unmodified).}

\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}

\item{verbose}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, show any warnings/messages per function.}
}
\value{
A data.frame with the columns \code{exported_function} and \code{code_file}.
}
\description{
Get all exported functions and map them to R script where they are defined
}
\keyword{internal}
