% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{map_risk}
\alias{map_risk}
\title{Use risk_breaks to map scores into character strings}
\usage{
map_risk(scores, risk_breaks, desc = FALSE)
}
\arguments{
\item{scores}{vector of risk scores (or other parameter if \code{desc = TRUE})}

\item{risk_breaks}{breaks determining low, medium, and high risk}

\item{desc}{Logical (T/F). If \code{TRUE}, sort \code{risk_breaks} in descending order.}
}
\description{
Use risk_breaks to map scores into character strings
}
\details{
Use \code{desc = TRUE} for color coding coverage and other parameters that follow and inverse relationship with risk.
}
\keyword{internal}
