% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-extra-notes.R
\name{map_tests_to_functions}
\alias{map_tests_to_functions}
\title{map test files and directories to all functions}
\usage{
map_tests_to_functions(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\description{
map test files and directories to all functions
}
\keyword{internal}
