% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{pkg_riskmetric}
\alias{pkg_riskmetric}
\title{Run all relevant riskmetric checks}
\usage{
pkg_riskmetric(pkg_path)
}
\arguments{
\item{pkg_path}{path to installed/untarred package}
}
\value{
raw riskmetric outputs
}
\description{
Run all relevant riskmetric checks
}
