% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{render_scorecard}
\alias{render_scorecard}
\title{Take a JSON from score_pkg() and render a pdf}
\usage{
render_scorecard(
  json_path,
  risk_breaks = c(0.3, 0.7),
  mitigation = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{json_path}{Path to a JSON file created by \code{\link[=score_pkg]{score_pkg()}}}

\item{risk_breaks}{A numeric vector of length 2, with both numbers being
between 0 and 1. These are used for the "breaks" when classifying scores
into Risk categories. For example, for the default value of \code{c(0.3, 0.7)},
for a given score: \verb{0 <= score < 0.3} is "Low Risk", \verb{0.3 <= score < 0.7}
is "Medium Risk", and \verb{0.7 <= score < 1} is "High Risk".}

\item{mitigation}{(Optional) path to a plain text file with any explanation
necessary for justifying use of a potentially "high risk" package.}
}
\description{
Take a JSON from score_pkg() and render a pdf
}
