% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard-summary.R
\name{render_scorecard_summary}
\alias{render_scorecard_summary}
\title{Render PDF summary of scorecards}
\usage{
render_scorecard_summary(json_paths, out_dir, snapshot = NULL)
}
\arguments{
\item{json_paths}{a vector of json paths}

\item{snapshot}{a report subtitle indicating the grouping of these packages, such as an MPN snapshot.}
}
\description{
Render PDF summary of scorecards
}
