% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{source_pkg_code}
\alias{source_pkg_code}
\title{Source R files into environment with encoding}
\usage{
source_pkg_code(files, file_encoding = "unknown", envir, verbose = FALSE)
}
\arguments{
\item{files}{vector of files to source}

\item{file_encoding}{encoding to be assumed for input strings.}

\item{envir}{an environment to store the sourced functions}

\item{verbose}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, show any warnings/messages per function.}
}
\description{
Source R files into environment with encoding
}
\details{
Inspired from pkgload internal functions: source_one, source_many and read_lines_enc
}
\keyword{internal}
