% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unpack_tarball}
\alias{unpack_tarball}
\title{Untar package and return installation directory}
\usage{
unpack_tarball(pkg_tar, temp_file_name = "SCORECARD_")
}
\arguments{
\item{pkg_tar}{path to tarball package}

\item{temp_file_name}{name of \code{tempfile}}
}
\description{
Untar package and return installation directory
}
\keyword{internal}
