% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{wrap_text}
\alias{wrap_text}
\title{Wrap a vector of strings using specific characters}
\usage{
wrap_text(
  str,
  width = 23,
  wrap_sym = c("/", "_", "-"),
  strict = FALSE,
  indent = FALSE
)
}
\arguments{
\item{str}{a string or vector of strings.}

\item{width}{Positive integer giving the target line width (in number of
characters).}

\item{wrap_sym}{Characters to loop through and attempt to make new lines with.
Note that the order you specify these values matters.}

\item{strict}{logical (T/F). If \code{FALSE}, will soft wrap based on the
characters specified via \code{wrap_sym}. If \code{TRUE}, will first soft wrap, and then
enforce the specified \code{width}.}

\item{indent}{logical (T/F). If \code{TRUE}, indent new lines via a \code{markdown} tab
(\code{'\\t'})}
}
\description{
Wrap a vector of strings using specific characters
}
\details{
\code{stringr::str_wrap} is not strict with the width for word characters, so
a helper function was needed. Rather than being 100\% strict at following the
cutoff width, we attempt to split at white space and specific symbols commonly
used in function names and file paths. If this splitting is not sufficient,
we then perform a strict operation, whereby we cut the line off at exactly
that width.
}
\keyword{internal}
