% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{add_coverage}
\alias{add_coverage}
\title{Run covr and potentially save results to disk}
\usage{
add_coverage(pkg_source_path, out_dir, timeout = Inf)
}
\arguments{
\item{pkg_source_path}{package installation directory}

\item{out_dir}{directory for saving results}

\item{timeout}{Timeout to pass to \code{\link[callr:r]{callr::r_safe()}} when running covr.}
}
\description{
Run covr and potentially save results to disk
}
\details{
The basename of \code{out_dir} should be the package name and version pasted together
}
\keyword{internal}
