% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{add_rcmdcheck}
\alias{add_rcmdcheck}
\title{Run R CMD CHECK}
\usage{
add_rcmdcheck(source_dir, out_dir, use_lib = .libPaths())
}
\arguments{
\item{source_dir}{package installation directory}

\item{out_dir}{directory for saving results}

\item{use_lib}{library path. Defaults to \code{.libPaths()}}
}
\description{
Run R CMD CHECK
}
\details{
rcmdcheck takes either a tarball or an installation directory
}
\keyword{internal}
