% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{add_rcmdcheck}
\alias{add_rcmdcheck}
\title{Run R CMD CHECK}
\usage{
add_rcmdcheck(out_dir, rcmdcheck_args)
}
\arguments{
\item{out_dir}{directory for saving results}

\item{rcmdcheck_args}{list of arguments to pass to \code{rcmdcheck::rcmdcheck}}
}
\description{
Run R CMD CHECK
}
\details{
rcmdcheck takes either a tarball or an installation directory.

The basename of \code{out_dir} should be the package name and version pasted together

The returned score is calculated via a weighted sum of notes (0.10), warnings (0.25), and errors (1.0). It has a maximum score of 1 (indicating no errors, notes or warnings)
and a minimum score of 0 (equal to 1 error, 4 warnings, or 10 notes). This scoring methodology is taken directly from \code{\link[riskmetric:metric_score.pkg_metric_r_cmd_check]{riskmetric::metric_score.pkg_metric_r_cmd_check()}}.
}
\keyword{internal}
