% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{add_rcmdcheck}
\alias{add_rcmdcheck}
\title{Run R CMD CHECK}
\usage{
add_rcmdcheck(pkg_path, out_dir, rcmdcheck_args)
}
\arguments{
\item{pkg_path}{package installation directory}

\item{out_dir}{directory for saving results}
}
\description{
Run R CMD CHECK
}
\details{
rcmdcheck takes either a tarball or an installation directory
}
\keyword{internal}
