% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard-summary.R
\name{build_risk_summary}
\alias{build_risk_summary}
\title{Summarise overall risks for each package}
\usage{
build_risk_summary(json_paths, risk_breaks, out_dir)
}
\arguments{
\item{json_paths}{A vector of json paths}

\item{risk_breaks}{A numeric vector of length 2, with both numbers being
between 0 and 1. These are used for the "breaks" when classifying scores
into Risk categories. For example, for the default value of \code{c(0.3, 0.7)},
for a given score: \verb{0 <= score < 0.3} is "Low Risk", \verb{0.3 <= score < 0.7}
is "Medium Risk", and \verb{0.7 <= score < 1} is "High Risk".}

\item{out_dir}{Output directory for saving scorecard summary. If \code{NULL}, assumes all \code{json_paths} point to the same output directory}
}
\description{
Summarise overall risks for each package
}
\keyword{internal}
