% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_exists_and_overwrite}
\alias{check_exists_and_overwrite}
\title{Check if a path exists and delete the file
if overwrite is TRUE}
\usage{
check_exists_and_overwrite(path, overwrite)
}
\description{
Check if a path exists and delete the file
if overwrite is TRUE
}
