% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{check_for_comments}
\alias{check_for_comments}
\title{Look for comment file and return contents if is found}
\usage{
check_for_comments(results_dir)
}
\arguments{
\item{results_dir}{directory containing json file and individual results. Output file path from \code{\link[=score_pkg]{score_pkg()}}}
}
\description{
The presence of a comment section indicates that we are aware the score is
low, but we are proceeding with adding the package to MPN or otherwise
approving it. This is not the same as a deviation from expected results.
}
\keyword{internal}
