% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{filter_symbol_functions}
\alias{filter_symbol_functions}
\title{Remove specific symbols from vector of functions}
\usage{
filter_symbol_functions(funcs)
}
\arguments{
\item{funcs}{vector of functions to filter}
}
\description{
Remove specific symbols from vector of functions
}
\keyword{internal}
