% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{find_export_script}
\alias{find_export_script}
\title{Find the R script containing the export}
\usage{
find_export_script(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\description{
We need to know the script so we can determine documentation percentage per script
}
\keyword{internal}
