% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{flex_stripe}
\alias{flex_stripe}
\title{Helper to apply flextable styling for alternating row background color}
\usage{
flex_stripe(
  tab,
  odd_body_bg = "#EFEFEF",
  even_body_bg = "transparent",
  border = TRUE
)
}
\arguments{
\item{tab}{a flextable}

\item{odd_body_bg, even_body_bg}{the background color of odd and even rows for
the main body of \code{tab}. Note that all font is black for the body.}

\item{border}{Logical (T/F). If \code{TRUE}, add horizontal borders and a vertical
border after the first column.}
}
\description{
Helper to apply flextable styling for alternating row background color
}
\keyword{internal}
