% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_extra_notes}
\alias{format_extra_notes}
\title{Format extra notes}
\usage{
format_extra_notes(extra_notes_data)
}
\arguments{
\item{extra_notes_data}{named list. Output of \code{\link[=create_extra_notes]{create_extra_notes()}}}
}
\description{
Format extra notes
}
\keyword{internal}
