% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_package_details}
\alias{format_package_details}
\title{Format Package Details}
\usage{
format_package_details(formatted_pkg_scores, color_headers = TRUE)
}
\arguments{
\item{formatted_pkg_scores}{list containing all scores}

\item{color_headers}{Logical (T/F). If \code{TRUE}, color the headers based on risk level}
}
\value{
a formatted flextable object
}
\description{
Formats Documentation, Maintenance and Transparency scores
}
\keyword{internal}
