% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_testing_scores}
\alias{format_testing_scores}
\title{Format testing scores}
\usage{
format_testing_scores(testing_df)
}
\arguments{
\item{testing_df}{dataframe returned by \code{formatted_pkg_scores$formatted$scores$testing}}
}
\value{
a formatted flextable object
}
\description{
Format testing scores
}
\keyword{internal}
