% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-report.R
\name{format_traceability_matrix}
\alias{format_traceability_matrix}
\title{Format Traceability Matrix}
\usage{
format_traceability_matrix(exports_df, wrap_cols = TRUE, scorecard_type = "R")
}
\arguments{
\item{exports_df}{tibble. Output of \code{\link[=make_traceability_matrix]{make_traceability_matrix()}}}

\item{wrap_cols}{logical (T/F). If \code{TRUE}, wrap columns}
}
\description{
Format Traceability Matrix
}
\keyword{internal}
