% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{get_all_functions}
\alias{get_all_functions}
\title{list all package functions}
\usage{
get_all_functions(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\description{
list all package functions
}
\keyword{internal}
