% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pkg_desc}
\alias{get_pkg_desc}
\title{Read Description file and parse the package name and version}
\usage{
get_pkg_desc(pkg_source_path, fields = NULL)
}
\arguments{
\item{pkg_source_path}{path to package source code (untarred)}
}
\description{
Read Description file and parse the package name and version
}
\keyword{internal}
