% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{get_toplevel_assignments}
\alias{get_toplevel_assignments}
\title{list all top-level objects defined in the package code}
\usage{
get_toplevel_assignments(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}
}
\value{
A data.frame with the columns \code{func} and \code{code_file} with a row for
every top-level object defined in the package.
}
\description{
This is primarily for getting all \emph{functions}, but it also returns top-level
declarations, regardless of type. This is intentional, because we also want
to capture any global variables or anything else that could be potentially
exported by the package.
}
\keyword{internal}
