% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-extra-notes.R
\name{make_traceability_matrix}
\alias{make_traceability_matrix}
\title{Create a Traceability Matrix}
\usage{
make_traceability_matrix(pkg_tar_path, results_dir = NULL, verbose = FALSE)
}
\arguments{
\item{pkg_tar_path}{path to a tarball}

\item{results_dir}{directory containing json file and individual results. Output file path from \code{\link[=score_pkg]{score_pkg()}}}

\item{verbose}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, show any warnings/messages per function.}
}
\value{
a tibble
}
\description{
Returns a table that links all exported functions and their aliases to their documentation (\code{man} files),
the R scripts containing them, and the test scripts that reference them.
}
\keyword{internal}
