% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{map_answer}
\alias{map_answer}
\title{Use answer_breaks to map binary results into character strings}
\usage{
map_answer(results, answer_breaks = c(0, 0.5, 1))
}
\arguments{
\item{answer_breaks}{breaks determining 'Yes' or 'No'}

\item{scores}{vector of risk scores (or other parameter if \code{desc = TRUE})}
}
\description{
Use answer_breaks to map binary results into character strings
}
\details{
If value is not found in \code{answer_breaks}, it is skipped over

Note: A result of 0.5 indicates an answer of 'Yes' with warnings. This is a special case for rmdcheck
}
\keyword{internal}
