% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-traceability-matrix.R
\name{map_functions_to_docs}
\alias{map_functions_to_docs}
\title{Map all Rd files to the functions they describe}
\usage{
map_functions_to_docs(pkg_source_path, verbose)
}
\arguments{
\item{pkg_source_path}{a file path pointing to an unpacked/untarred package directory}

\item{verbose}{Logical (\code{TRUE}/\code{FALSE}). If \code{TRUE}, show any warnings/messages per function.}
}
\description{
Map all Rd files to the functions they describe
}
\keyword{internal}
