% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{map_risk}
\alias{map_risk}
\title{Use risk_breaks to map scores into character strings}
\usage{
map_risk(scores, risk_breaks)
}
\arguments{
\item{scores}{vector of risk scores}

\item{risk_breaks}{breaks determining low, medium, and high risk}
}
\description{
Use risk_breaks to map scores into character strings
}
\keyword{internal}
