% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskmetric.R
\name{pkg_riskmetric}
\alias{pkg_riskmetric}
\title{Run all relevant riskmetric checks}
\usage{
pkg_riskmetric(pkg_source_path)
}
\arguments{
\item{pkg_source_path}{path to package source code (untarred)}
}
\value{
raw riskmetric outputs
}
\description{
Run all relevant riskmetric checks
}
