% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard.R
\name{render_scorecard}
\alias{render_scorecard}
\title{Render a scorecard PDF from a directory of results}
\usage{
render_scorecard(
  results_dir,
  risk_breaks = c(0.3, 0.7),
  overwrite = FALSE,
  add_traceability = "auto"
)
}
\arguments{
\item{results_dir}{Directory with scoring results. This is the path returned
by \code{\link[=score_pkg]{score_pkg()}}.}

\item{risk_breaks}{A numeric vector of length 2, with both numbers being
between 0 and 1. These are used for the "breaks" when classifying scores
into Risk categories. For example, for the default value of \code{c(0.3, 0.7)},
for a given score: \verb{0 <= score < 0.3} is "Low Risk", \verb{0.3 <= score < 0.7}
is "Medium Risk", and \verb{0.7 <= score < 1} is "High Risk".}

\item{overwrite}{Logical (T/F). If \code{TRUE}, will overwrite an existing file path if it exists}

\item{add_traceability}{Logical (T/F). If \code{TRUE}, append a table that links package functionality to the documentation and test files.
Defaults to "auto", which will include the matrix if found.}
}
\description{
Create a scorecard from a results directory prepared by \code{\link[=score_pkg]{score_pkg()}} or an
external scorer (see \link{external_scores}).
}
\details{
If a plain text comments file is found in \code{results_dir}, it will automatically be included.
\strong{Note} that it must follow the naming convention of \verb{<pkg_name>_<pkg_version>.comments.txt}

If a traceability matrix is found in \code{results_dir}, it will automatically be included unless overridden via \code{add_traceability}.
}
