% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard-summary.R
\name{render_scorecard_summary}
\alias{render_scorecard_summary}
\title{Render PDF summary of scorecards}
\usage{
render_scorecard_summary(
  json_paths,
  risk_breaks = c(0.3, 0.7),
  out_dir = NULL,
  snapshot = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{json_paths}{A vector of json paths}

\item{risk_breaks}{A numeric vector of length 2, with both numbers being
between 0 and 1. These are used for the "breaks" when classifying scores
into Risk categories. For example, for the default value of \code{c(0.3, 0.7)},
for a given score: \verb{0 <= score < 0.3} is "Low Risk", \verb{0.3 <= score < 0.7}
is "Medium Risk", and \verb{0.7 <= score < 1} is "High Risk".}

\item{out_dir}{Output directory for saving scorecard summary. If \code{NULL}, assumes all \code{json_paths} point to the same output directory}

\item{snapshot}{A report subtitle indicating the grouping of these packages, such as an MPN snapshot. See details.}

\item{overwrite}{Logical (T/F). If \code{TRUE}, will overwrite an existing file path if it exists}
}
\description{
Render PDF summary of scorecards
}
\details{
The argument \code{snapshot}, will appear in the document with language similar to '\strong{MPN Snapshot \verb{<snapshot>}}'.
In other words, \code{snapshot} should be a date, or some other grouping that makes sense grammatically according to the this example.
}
