% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-pkg.R
\name{score_pkg}
\alias{score_pkg}
\title{Generate scorecard metrics for package}
\usage{
score_pkg(pkg, out_dir, pkg_info = NULL)
}
\description{
Returns collected metrics as a named list, and also writes
results to disk for downstream consumption.
}
