% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-scorecard-summary.R
\name{summarize_package_results}
\alias{summarize_package_results}
\title{Review important package metrics prior to running \code{\link[=render_scorecard_summary]{render_scorecard_summary()}}}
\usage{
summarize_package_results(result_dirs)
}
\arguments{
\item{result_dirs}{A vector of output directories}
}
\value{
a dataframe
}
\description{
Review important package metrics prior to running \code{\link[=render_scorecard_summary]{render_scorecard_summary()}}
}
